Use NaN as default resolution for TimeSeries objects
The NWB schema documentation says

Smallest meaningful difference between values in data, stored in the specified by unit.
Comment: E.g., the change in value of the least significant bit, or a
larger number if signal noise is known to be present. If unknown, use NaN.

So using NaN instead of zero is a more natural choice as a resolution of zero
means that you have perfect experimental conditions which are in general
rarely available.