PEP8Speaks: Only scan the diff of the changed files for violations
Currently PEP8Speaks checks all changed files. This results in warning
messages for lines which were not touched in a given PR.

This is, at least for casual contributors, quite annoying as it basically
forces them to cleanup after other people's contributions.

Let's scan only the changed lines to make it less annoying.