add datetime support
Specification for `file_create_date` and `session_start_time` are
updated to only accept iso8601 date format.

since hdf5 does not accept np.datetime64 and datetime64 is
time-zone-naive, a new dtype `isodatetime` is introduced that
transformes datetime objects to string representation on write.

on read, the string has to be transformed back. The builder
representation has the bare hdf5object. The Main File Container Class
now only accepts datetime objects. Therefore, the transformation between
string and datetime is done in the NWBFileMap definition as an override
function.

The File Class adds the local timezone object if no object was given as
input.

date conversion from datetime to iso string using datetime.isoformat()
conversion back from iso string to datetime is done using the parse
function fo the dateutils library.  `datetime.fromisoformat` would be
better here but is only available in python3.7 and is missing in
python2.7 (2018-10-03)

description of session_start_time and file_create_date was updated to
allow UTC timezone ("Z") and an accuracy up to milliseconds (note that
np.datetime64 uses nanoseconds)

unit tests with string inputs were updated to get datetime objects.

a function is added for converting datetime objects with missing
timezones.  A warning is raised when a datetime object is used with
missing timezone information.