Add a SweepTable for icephys data
For icephys a common data acquisition method is to acquire data from
multiple electrodes in one run. This "run" is commonly called a sweep.

Now at the moment there exists no way of grouping PatchClampSeries
together in a NWB file.

The SweepTable solves that problem. It is a table with two columns
"sweep_number" and a vector of links to PatchClampSeries.

Now whenever someone adds some TimeSeries via
NWBFile.add_acquisition/add_stimulus/add_stimulus_template we check if
the added element is a PatchClampSeries and has a valid sweep number. In
that case we add, or append, to the SweepTable.

The sweep table is located at `/general/intracellular_ephys/sweep_table`
and is a DynamicTable.