Add test for validation cli (#1164)
* test.py/validate_nwbs: Increment TOTAL in the loop

This prepares for a future commit which will have a complex to guess
number of runs and therefore incrementing it on every iteration is more
feasible.

* test.py/validate: Add validation by invoking the CLI interface

As previous bug reports have shown the CLI interface for validation
works differently due to namespace caching.

In order to catch these problems earlier on, we now validate the files
using python -m pynwb.validate as well using the following variations
- cached namespace info
- non-cached namespace info
- single namespace
- all namespaces

* src/pynwb/validate.py: Add list of available namespaces in the error case for --ns

* src/pynwb/validate.py: Add CLI option to output the available namespaces

* src/pynwb/validate.py: Fix missing space in outputted message

* docs/source/validation.rst: Rewrite and enhance

* Fix validating NWB files against non-cached schemas

Use 'core' namespace as default instead of available_namespaces().
available_namespaces() returns ['core', 'hdmf-common']. However, PyNWB
types will not validate against the HDMF namespace because the types are
missing the data type key 'data_type' that is necessary for validation.
The PyNWB core schema includes hdmf.common, so validating against the
pynwb namespace will also validate the hdmf.common types. For example,
when a VectorData is written, it is written with attributes
'namespace=core' and 'neurodata_type=VectorData' without mention of
hdmf. So, we need only validate files against the PyNWB core namespace.

Co-authored-by: Ryan Ly <rly@lbl.gov>